Created EntriesDAO and amended NotesDAO such that EntriesDAO and NotesDAO depend on EntriesDAO.
